[M]shell优化

+++ home/start.sh
@@ -1,7 +1,10 @@
+if [ -s `which lsof`];
+then
 sed -i 's#http://deb.debian.org#https://mirrors.ustc.edu.cn#g' /etc/apt/sources.list
 sed -i 's|security.debian.org/debian-security|mirrors.ustc.edu.cn/debian-security|g' /etc/apt/sources.list
 
 apt update && apt install lsof
+fi
 
 cd /root/apache-atlas-2.3.0
